Add framework for landing page
Modify .gitignore to include bundle.js and jsconfig.js. bundle.js is
a compiled file, and jsconfig.js is a local config file.

Modify package.json to add dependencies for babel, react, react-dom,
and webpack.

Add index.html and index.js as basic starting points for the landing
page and the rest of the react app.

Add webpack.config.js as configuration file for webpack. it targets
index.js and compiles it and all it's dependencies using babel. The
output is recorded in bundle.js. This file can be sent as a single self
contained file to the client to render the app.