Create Express Server, Routes and Implement Login
Use express to serve UI and routes for UI. Try and implement ajax request
to api. The emphasis was to minimize the use of render as much of the UI
in React as possible

Still need to deal with session cookie!
Should deal with Flash Messages in React

To run server:
1 npm install

2 npm start