Add login.js for the login link
login.js is a component that contains the login/logout link that will
sit in the menu bar. It request authorization strategies from the API
and then uses that strategy to initialize the authentication process by
linking to the third party Oauth API. It also Logs the user out by
deleting the jwt from local storage and redirecting the user to the
home page.

Add readFromAPI() to the root component so that any child compent can
call it if it needs to make an API call. readFromAPI(url, callback)
handles get requests and passes the return values as a paramater in the
callback.

Add more dependencies to package.json