Add Redux Store Functionality
Update index.html to so root div is labled root not app

start up configured store with async functions in index.js!

Add new dependencies to package.json

I have added functionality for getting curriculum and getting auth
strategies to redux, I have also added fuctionality for showing and
hiding a redirect modal to the redux store. I had to modify the code in
a couple of components to make them work with the redux store.