Restyle for repsonsive UI on mobile. Remove 'localhost'.
Removed localhost from server.js so page can be accessed by other devices
on the network, allowing easier testing. Removed Semantic UI in favor or
React-MDL. This is a point that needs discussion, but I think it leads to
a cleaner UI. And easier development. Refactor components to reflect this
change. Components are still WIP.